refactor: pull data transfer object code out of database
